requirements.txt
requests
discord.py
flask
requests
loguru
pony
terminaltables
diskcache
fire
faker
ipython
toml
